backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-dotenv==1.1.1
Pillow==11.3.0
django-filter==25.1
django-extensions==4.1
celery==5.5.3
redis==6.4.0
openai==1.106.1
pandas==2.3.2
numpy==2.2.6
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.0
supabase==2.18.1
dj-database-url==3.0.1
